ccg015 fix remove+ copyTopPage css +focus item when close dialog

+++ nts.uk/ctx/sys/portal/nts.uk.ctx.sys.portal.dom/src/main/java/nts/uk/ctx/sys/portal/dom/toppage/service/internal/TopPageServiceImpl.java
@@ -63,9 +63,11 @@ public class TopPageServiceImpl implements TopPageService {
 	public void removeTopPage(String topPageCode, String companyId) {
 		TopPage tp = topPageRepository.findByCode(companyId, topPageCode).get();
 		topPageRepository.remove(companyId, topPageCode);
-		if (!tp.getLayoutId().equals("")) {
-			//TODO change to service of Lam san
-			layoutRepository.remove(companyId, tp.getLayoutId());
+		if (tp.getLayoutId() != null) {
+			if (!tp.getLayoutId().equals("")) {
+				// TODO change to service of Lam san
+				layoutRepository.remove(companyId, tp.getLayoutId());
+			}
 		}
 	}
 }
